ADM-733[backend]refactor: rename AsyncReportDataHandler to AsyncDataBaseHandler
