+++ backend/src/main/java/heartbeat/handler/AsyncExceptionHandler.java
@@ -2,6 +2,7 @@ package heartbeat.handler;
 
 import com.google.gson.Gson;
 import heartbeat.exception.BaseException;
+import heartbeat.handler.base.AsyncDataBaseHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Component;
@@ -11,7 +12,7 @@ import static heartbeat.handler.FIleType.ERROR;
 @Log4j2
 @Component
 @RequiredArgsConstructor
-public class AsyncExceptionHandler extends AsyncReportDataHandler {
+public class AsyncExceptionHandler extends AsyncDataBaseHandler {
 
 	public void put(String reportId, BaseException e) {
 		createDirToConvertData(ERROR);
